Move write_atomic to util module
